Fixed memory bug in assignMacros.cpp relating to ROOT file opening/closure; also fixed accidental bleed-over of macroTime waveform data into detector waveforms.

+++ analysis/include/raw.h
@@ -12,7 +12,7 @@ bool readWaveformEventBody(std::ifstream& evtfile);
 double calculateCFDTime(std::vector<int>* waveform,
         double baseline,
         double fraction,
-        double delay,
+        unsigned int delay,
         bool isPositiveSignal);
 
 #endif
